Update the UI for water intake

+++ ChartsExample/Chart/WaterIntakeChart/WaterIntakeChartDataEntry.swift
@@ -10,12 +10,12 @@ import UIKit
 import Charts
 
 class WaterIntakeChartDataEntry: BubbleChartDataEntry {
-    var count: Int = 0
+    var count: Double = 0
     var colors: [UIColor] = [UIColor.white] // gradient color from top to bottom
     var highlightedMultipler: CGFloat = 1.2
     var normalizedHeight: Double = 0
     
-    init(x: Double, y: Double, size: CGFloat, count: Int) {
+    init(x: Double, y: Double, size: CGFloat, count: Double) {
         super.init(x: x, y: y, size: size)
         self.normalizedHeight = Double(size)
         self.count = count
